+++ common/src/main/java/com/joyent/http/signature/crypto/NativeSHA256.java
@@ -1,9 +1,11 @@
 package com.joyent.http.signature.crypto;
 
+import com.joyent.http.signature.CryptoException;
 import org.bouncycastle.asn1.nist.NISTObjectIdentifiers;
-import org.bouncycastle.crypto.digests.SHA256Digest;
+import org.bouncycastle.crypto.Digest;
 import org.bouncycastle.crypto.encodings.PKCS1Encoding;
 import org.bouncycastle.jcajce.provider.asymmetric.rsa.DigestSignatureSpi;
+import org.bouncycastle.jcajce.provider.util.DigestFactory;
 
 /**
  * Simple wrapper class for providing SHA256 RSA signing using native libraries.
@@ -16,7 +18,22 @@ public class NativeSHA256 extends DigestSignatureSpi {
      */
     public NativeSHA256() {
         super(NISTObjectIdentifiers.id_sha256,
-              new SHA256Digest(),
+              getDigest("SHA-256"),
               new PKCS1Encoding(new NativeRSABlindedEngine()));
+    }
+
+    /**
+     * Finds checksum by name.
+     *
+     * @param digest digest name
+     * @return digest instance
+     */
+    private static Digest getDigest(final String digest) {
+        try {
+            return new JceDigest(digest);
+        } catch (CryptoException e) {
+            String noHyphen = digest.replaceFirst("-", "");
+            return DigestFactory.getDigest(noHyphen);
+        }
     }
 }